Renamed router dir
